deps: bump criterion, tree-sitter, tree-sitter-bash
- criterion 0.5 -> 0.8: replace deprecated criterion::black_box with std::hint::black_box
- tree-sitter 0.24 -> 0.26 / tree-sitter-bash 0.23 -> 0.25: Node::child now takes u32

+++ crates/lib/src/tools/bash_parse.rs
@@ -55,7 +55,7 @@ fn extract_from_node(node: Node, source: &[u8], parsed: &mut ParsedCommand) {
             } else {
                 // Fallback: first word child.
                 for i in 0..node.child_count() {
-                    let child = node.child(i).unwrap();
+                    let child = node.child(i as u32).unwrap();
                     if child.kind() == "word" || child.kind() == "command_name" {
                         parsed.commands.push(node_text(child, source));
                         break;
@@ -86,7 +86,7 @@ fn extract_from_node(node: Node, source: &[u8], parsed: &mut ParsedCommand) {
             parsed.has_pipes = true;
             // Extract each command in the pipeline.
             for i in 0..node.child_count() {
-                let child = node.child(i).unwrap();
+                let child = node.child(i as u32).unwrap();
                 if child.kind() == "command" || child.kind() == "pipeline" {
                     let text = node_text(child, source);
                     parsed.pipeline_segments.push(text);
@@ -96,7 +96,7 @@ fn extract_from_node(node: Node, source: &[u8], parsed: &mut ParsedCommand) {
         "list" => {
             // && and || chains.
             for i in 0..node.child_count() {
-                let child = node.child(i).unwrap();
+                let child = node.child(i as u32).unwrap();
                 let kind = child.kind();
                 if kind == "&&" || kind == "||" {
                     parsed.has_chains = true;
@@ -115,7 +115,7 @@ fn extract_from_node(node: Node, source: &[u8], parsed: &mut ParsedCommand) {
 
     // Recurse into children.
     for i in 0..node.child_count() {
-        if let Some(child) = node.child(i) {
+        if let Some(child) = node.child(i as u32) {
             extract_from_node(child, source, parsed);
         }
     }
